Add arg (-q, --quality): increases passhprase quality by increasing its entropy
This is useful in instances where the user may not want to type long passphrases,
yet nevertheless have a strong password. Also useful in instances where the target system
places length limits on the password. When this argument is used the program will
insert one special character at random into the string and change one letter at random
to uppercase.